+++ game/content/ghplots/mysterymission.py
@@ -1,4 +1,5 @@
 from typing import override
+from game.content.ghplots import missionbuilder
 from pbge.plots import Plot
 import game
 import gears
@@ -24,6 +25,142 @@ from game.content.ghplots.missionbuilder import MAIN_OBJECTIVE_VALUE
 # A mystery mission may add an ongoing side story, a permanent complication to the campaign, 
 # or who knows what else!
 BAMO_MYSTERYMISSION = "BAMO_MYSTERYMISSION"
+
+
+class MyMi_DoubleEncounter(Plot):
+    LABEL = BAMO_MYSTERYMISSION
+    LABEL = missionbuilder.BAMO_TEST_MISSION
+    active = True
+    scope = "LOCALE"
+
+    GOOD_ENEMIES = (
+        gears.factions.BioCorp, gears.factions.KettelIndustries, gears.factions.RegExCorporation,
+        gears.factions.Ravagers, gears.factions.BladesOfCrihna, gears.factions.BoneDevils,
+        gears.factions.TreasureHunters, gears.factions.ProDuelistAssociation
+    )
+
+    def custom_init(self, nart):
+        myscene = self.elements["LOCALE"]
+        a, b = random.sample(self.GOOD_ENEMIES, 2)
+        myfac = self.elements.setdefault("ENEMY_FACTION", a)
+        roomtype = self.elements["ARCHITECTURE"].get_a_room()
+        self.register_element("ROOM", roomtype(15, 15), dident="LOCALE")
+
+        team2 = self.register_element("_eteam", teams.Team(enemies=(myscene.player_team,)), dident="ROOM")
+        team3 = self.register_element("_bteam", teams.Team(enemies=(myscene.player_team, team2)))
+
+        mynpc = self.register_element("NPC", nart.camp.cast_a_combatant(
+            myfac, self.rank, opposed_faction=self.elements["ALLIED_FACTION"], myplot=self
+        ), lock=True)
+        _=plotutility.CharacterMover(nart.camp, self, mynpc, myscene, team2)
+        myunit = gears.selector.RandomMechaUnit(self.rank, 120, myfac, myscene.environment, add_commander=False)
+
+        team2.contents += myunit.mecha_list
+
+        myunit = gears.selector.RandomMechaUnit(self.rank, 200, b, myscene.environment, add_commander=True)
+        self.elements["_commander2"] = myunit.commander
+        team3.contents += myunit.mecha_list
+
+        self.obj = adventureseed.MissionObjective("Investigate mecha activity nearby",
+                                                  MAIN_OBJECTIVE_VALUE * 2)
+        self.adv.objectives.append(self.obj)
+
+        self.intro_ready = True
+        self.team_b_deployed = False
+
+        team4 = self.register_element("_cargoteam", teams.Team(), dident="ROOM")
+        team4.contents.append(game.content.plotutility.CargoContainer())
+
+        tgen = gears.artifacts.ArtifactBuilder(self.rank)
+        self.elements["TREASURE"] = tgen.item
+
+        return True
+
+    def _eteam_ACTIVATETEAM(self, camp):
+        if self.intro_ready:
+            npc = self.elements["NPC"]
+            ghdialogue.start_conversation(camp, camp.pc, npc, cue=ghdialogue.ATTACK_STARTER)
+            self.intro_ready = False
+
+    def t_ENDCOMBAT(self, camp: gears.GearHeadCampaign):
+        ateam: teams.Team = self.elements["_eteam"]
+        bteam: teams.Team = self.elements["_bteam"]
+        cteam: teams.Team = self.elements["_cargoteam"]
+
+        if len(ateam.get_members_in_play(camp)) < 1:
+            if not self.team_b_deployed:
+                _=pbge.alert("Just when you think the battle is over, a second force arrives to claim the cargo.")
+                camp.scene.deploy_team(bteam.contents, bteam, self.elements["ROOM"].area)
+                self.team_b_deployed = True
+            elif len(bteam.get_members_in_play(camp)) < 1:
+                if len(cteam.get_members_in_play(camp)) > 0:
+                    _=plotutility.ItemGiverWithDisplay(camp, self.elements["TREASURE"], "After the battle, you discover a {} in the cargo container.")
+                    self.obj.win(camp, 100)
+                else:
+                    self.obj.win(camp, 50)
+
+    def NPC_offers(self, camp: gears.GearHeadCampaign):
+        mylist = list()
+        mylist.append(Offer("[WHATAREYOUDOINGHERE] I was supposed to recover this cargo before anyone else arrives...",
+                            context=ContextTag([context.ATTACK, ])))
+
+        mylist.append(Offer(
+            "[SO_IT_SEEMS] Maybe we can finish before anyone else arrives. [LETSFIGHT]",
+            context=ContextTag([context.COMBAT_CUSTOM, ]),
+            data={"reply": "Looks like you're too late for that."}, effect=self.friendly_battle
+        ))
+
+        _=game.ghdialogue.SkillBasedPartyReply(
+            Offer(
+                "[GOOD_POINT] In that case, I'll leave you to it... my sensors indicate you won't have to wait long.",
+                context=ContextTag([context.COMBAT_CUSTOM]), effect=self.friendly_retreat,
+                data={"reply": "Then you don't stand a chance against both us and whoever else is looking."}
+            ), camp, mylist, gears.stats.Ego, gears.stats.Negotiation, self.rank,
+            difficulty=gears.stats.DIFFICULTY_HARD, no_random=False
+        )
+
+        if not camp.is_unfavorable_to_pc(self.elements["NPC"]):
+            mylist.append(Offer(
+                "[THANK_YOU] [GOODBYE]",
+                context=ContextTag([context.COMBAT_CUSTOM, ]),
+                data={"reply": "No worries. I'll get out of your way, then."},
+                effect=self.friendly_withdraw
+            ))
+
+        return mylist
+
+    def friendly_battle(self, camp: gears.GearHeadCampaign):
+        npc: gears.base.Character = self.elements["NPC"]
+        npc.relationship = camp.get_relationship(npc)
+        if not npc.relationship.role and not camp.is_favorable_to_pc(npc.faction):
+            npc.relationship.role = gears.relationships.R_OPPONENT
+        npc.relationship.history.append(gears.relationships.Memory(
+            "we fought over a mysterious cargo container",
+            "I fought you over a mysterious cargo container",
+            5, memtags=(gears.relationships.MEM_Clash, gears.relationships.MEM_Ideological)
+        ))
+
+    def friendly_retreat(self, camp):
+        npc: gears.base.Character = self.elements["NPC"]
+        npc.relationship = camp.get_relationship(npc)
+        npc.relationship.history.append(gears.relationships.Memory(
+            "I left you to fight whoever for a cargo container",
+            "I let you go home from a mission early without even fighting",
+            10, memtags=(gears.relationships.MEM_Clash, gears.relationships.MEM_CallItADraw)
+        ))
+        self.elements["_eteam"].retreat(camp)
+
+    def friendly_withdraw(self, camp):
+        npc: gears.base.Character = self.elements["NPC"]
+        npc.relationship = camp.get_relationship(npc)
+        npc.relationship.history.append(gears.relationships.Memory(
+            "you let me finish my mission without fighting",
+            "I let you have that cargo container without putting up a fight",
+            10, memtags=(gears.relationships.MEM_Clash, gears.relationships.MEM_CallItADraw)
+        ))
+        self.obj.win(camp, 25 + npc.get_reaction_score(camp.pc, camp))
+        camp.scene.player_team.retreat(camp)
+
 
 
 class MyMi_LancemateTimedDefense(mission_bigobs.BAM_TimedDefense):
@@ -61,7 +198,7 @@ class MyMi_LancemateTimedDefense(mission_bigobs.BAM_TimedDefense):
         super().t_PCMOVE(camp)
         if self.combat_started and self.convo_ready:
             ghcutscene.SimpleMonologueDisplay(
-                "[HELP_ME_VS_MECHA_COMBAT] This [town] is under attack.",
+                "[HELP_ME_VS_MECHA_COMBAT] This [town] is under attack. We just need to hold on until reinforcements arrive.",
                 self.elements["NPC"].get_root()
             )(camp)