+++ game/exploration.py
@@ -677,7 +677,7 @@ class Explorer(object):
                     elif gdi.unicode == "@" and pbge.util.config.getboolean("GENERAL", "dev_mode_on"):
                         for thing in self.scene.contents:
                             if hasattr(thing, "pos"):
-                                print("{}: {}".format(thing, thing.pos))
+                                print("{}: {} {}".format(thing, thing.pos, getattr(thing, "mmode", "Undefined")))
 
                     elif gdi.unicode == "*" and pbge.util.config.getboolean("GENERAL", "dev_mode_on"):
                         for thing in self.camp.all_contents(self.camp):